# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PreventivoAI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /catalog:
    get:
      operationId: listCatalogItems
      tags: [catalog]
      summary: List price catalog items for the authenticated user
      responses:
        "200":
          description: List of catalog items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CatalogItem"
    post:
      operationId: createCatalogItem
      tags: [catalog]
      summary: Create a new price catalog item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCatalogItemBody"
      responses:
        "201":
          description: Created catalog item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogItem"
  /catalog/bulk:
    post:
      operationId: bulkCreateCatalogItems
      tags: [catalog]
      summary: Bulk create price catalog items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/CreateCatalogItemBody"
      responses:
        "201":
          description: Created catalog items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CatalogItem"
...
  /documents/{id}/extract:
    post:
      operationId: extractDocument
      tags: [documents]
      summary: Trigger AI extraction on an uploaded document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Extraction complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadedDocument"
...
  /clients/{id}/quotes:
    get:
      operationId: listClientQuotes
      tags: [clients]
      summary: List quotes for a specific client by stable id (md5 of normalized name)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of quotes for the client
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quote"
...
  /catalog/{id}:
    put:
      operationId: updateCatalogItem
      tags: [catalog]
      summary: Update a price catalog item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCatalogItemBody"
      responses:
        "200":
          description: Updated catalog item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogItem"
    delete:
      operationId: deleteCatalogItem
      tags: [catalog]
      summary: Delete a price catalog item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    QuoteItem:
      type: object
      properties:
        descrizione:
          type: string
        quantita:
          type: number
        unita:
          type: string
        prezzoUnitario:
          type: number
        totale:
          type: number
      required:
        - descrizione
        - quantita
        - unita
        - prezzoUnitario
        - totale
    QuoteChapterItem:
      type: object
      properties:
        descrizione:
          type: string
        um:
          type: string
        quantita:
          type: number
        prezzoUnitario:
          type: number
        totale:
          type: number
      required:
        - descrizione
        - um
        - quantita
        - prezzoUnitario
        - totale
    QuoteChapter:
      type: object
      properties:
        lettera:
          type: string
        titolo:
          type: string
        voci:
          type: array
          items:
            $ref: "#/components/schemas/QuoteChapterItem"
        subtotale:
          type: number
        osservazione:
          type: string
      required:
        - lettera
        - titolo
        - voci
        - subtotale
    QuoteDiscount:
      type: object
      properties:
        percentuale:
          type: number
        importoScontato:
          type: number
      required:
        - percentuale
        - importoScontato
    QuoteClientData:
      type: object
      properties:
        nome:
          type: string
        indirizzo:
          type: string
        email:
          type: string
        phone:
          type: string
        businessNumber:
          type: string
        partitaIva:
          type: string
        city:
          type: string
        postalCode:
          type: string
        province:
          type: string
      required:
        - nome
        - indirizzo
    QuoteCompanySnapshot:
      type: object
      properties:
        companyName:
          type: string
        vatNumber:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        logoUrl:
          type: string
      required:
        - companyName
    QuoteAttachment:
      type: object
      properties:
        id:
          type: string
        fileName:
          type: string
        mimeType:
          type: string
        fileUrl:
          type: string
        fileSize:
          type: number
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - fileName
        - mimeType
        - fileUrl
        - createdAt
    Quote:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        clientData:
          $ref: "#/components/schemas/QuoteClientData"
        descrizioneGenerale:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/QuoteItem"
        capitoli:
          type: array
          items:
            $ref: "#/components/schemas/QuoteChapter"
        sconto:
          $ref: "#/components/schemas/QuoteDiscount"
          nullable: true
        condizioniPagamento:
          type: array
          items:
            type: string
        titoloPreventivoRiga1:
          type: string
          nullable: true
        titoloPreventivoRiga2:
          type: string
          nullable: true
        numeroPreventivoData:
          type: string
          nullable: true
        companySnapshot:
          $ref: "#/components/schemas/QuoteCompanySnapshot"
          nullable: true
        subtotale:
          type: number
        ivaPercentuale:
          type: number
        ivaValore:
          type: number
        totale:
          type: number
        note:
          type: string
        status:
          type: string
          enum: [draft, unlocked, pending_payment]
        pdfUrl:
          type: string
          nullable: true
        rawInput:
          type: string
        pdfDownloadedAt:
          type: string
          nullable: true
        capitolatoPro:
          type: boolean
        capitolatoPdfUrl:
          type: ["string", "null"]
        templateId:
          type: string
          enum: [standard, arosio, mariagrazia]
          nullable: true
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/QuoteAttachment'
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - clientData
        - descrizioneGenerale
        - items
        - capitoli
        - condizioniPagamento
        - subtotale
        - ivaPercentuale
        - ivaValore
        - totale
        - note
        - status
        - rawInput
        - capitolatoPro
        - createdAt
        - updatedAt
...
    CatalogItem:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        nome:
          type: string
        categoria:
          type: string
          nullable: true
        um:
          type: string
        prezzoUnitario:
          type: number
        note:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - nome
        - um
        - prezzoUnitario
        - createdAt
        - updatedAt
    CreateCatalogItemBody:
      type: object
      properties:
        nome:
          type: string
        categoria:
          type: string
        um:
          type: string
        prezzoUnitario:
          type: number
        note:
          type: string
      required:
        - nome
        - um
        - prezzoUnitario
    UpdateCatalogItemBody:
      type: object
      properties:
        nome:
          type: string
        categoria:
          type: string
        um:
          type: string
        prezzoUnitario:
          type: number
        note:
          type: string
...
    UploadedDocument:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        fileName:
          type: string
        fileSize:
          type: integer
          nullable: true
        mimeType:
          type: string
        fileUrl:
          type: string
        status:
          type: string
          enum: [pending, processing, done, error]
        extractedData:
          $ref: "#/components/schemas/ExtractedDocumentData"
          nullable: true
        errorMessage:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - fileName
        - mimeType
        - fileUrl
        - status
        - createdAt
        - updatedAt
    ExtractedDocumentData:
      type: object
      properties:
        lavorazioni:
          type: array
          items:
            $ref: "#/components/schemas/ExtractedLavorazione"
          nullable: true
        totale:
          type: number
          nullable: true
        zona:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
    ExtractedLavorazione:
      type: object
      properties:
        tipo:
          type: string
        prezzoUnitario:
          type: number
        um:
          type: string
          nullable: true
        zona:
          type: string
          nullable: true
      required:
        - tipo
        - prezzoUnitario
...